Form controller to store
